修改db实体类的位置，generator中添加自定义方法和修改in方法

+++ common/general/src/main/java/com/cloud/mvc/example/general/supers/BaseEntity.java
@@ -8,8 +8,8 @@ import java.time.LocalDateTime;
 public class BaseEntity {
     private LocalDateTime createDate;
     private LocalDateTime modifyDate;
-    private Integer createUserId;
-    private Integer modifyUserId;
+    private Long createUserId;
+    private Long modifyUserId;
     private Integer status;
     private Integer version;
 }
